Execute composer.phar with the PHP binary running zf-deploy
If there is no PHP in the executable path then calling composer.phar as
executable fails. On Windows even a popup dialog appears for the user to
find the program to execute the phar.

This change resolves the issue by prepending the PHP binary path to the
command. However, this is not done in case 'composer' is available in
the executable path. In this case, we assume there is also PHP
available.